pyproject.toml
[tool.poetry]
name = "arc-to-dataverse"
version = "0.1.0"
description = ""
authors = ["Kevin Schneider <schneider.kev@outlook.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
jsonschema = "^4.23.0"
fsspreadsheet = "^6.2.0"
pydataverse = "^0.3.4"
arctrl = "^2.2.2"
python-dotenv = "^1.0.1"
typer = "^0.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
